+++ game/player.py
@@ -1,6 +1,3 @@
-from game.bag_tiles import BagTiles
-
-
 class Player:
     def __init__(self, bag_tiles, player_id):
         self.player_name = None
@@ -28,3 +25,35 @@ class Player:
             raise KeyError(e)
         
         return True
+    
+    def switch(self):
+        player_tiles = self.tiles
+        self.tiles = []
+        new_tiles = self.bag_tiles.take(7)
+        self.tiles.extend(new_tiles)
+        self.bag_tiles.put(player_tiles)
+        
+
+    # def switch(self, tiles_to_change):
+    #     if tiles_to_change < 1:
+    #         print("Debes cambiar al menos una ficha.")
+    #         return
+    #     if tiles_to_change > len(self.tiles):
+    #         print("No tienes suficientes fichas para cambiar esa cantidad.")
+    #         return
+
+    #     changes = input("Ingresa las letras de las fichas que deseas cambiar (sin espacios): ").upper()
+    #     changes = list(changes)
+
+    #     if len(changes) != tiles_to_change or not all(tile in [t.letter for t in self.tiles] for tile in changes):
+    #         print("Selección de fichas inválida.")
+    #         return
+
+    #     new_tiles = self.bag_tiles.take(tiles_to_change)
+    #     for tile in new_tiles:
+    #         self.tiles.remove(tile)
+
+    #     for tile_letter in changes:
+    #         self.tiles.append(self.bag_tiles.take(1, tile_letter)[0])
+
+    #     self.bag_tiles.put(new_tiles)